# storage/api-docs/api-docs-admin-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API",
        "description": "API documentation for the School Management System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/"
        }
    ],
    "paths": {
...
        "/admin/course": {
            "post": {
                "tags": [
                    "Course Management"
                ],
                "summary": "Create a New Course",
                "description": "Add a new course to the system.",
                "operationId": "3b426eb4e27697a962588fae7c2b537f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_name",
                                    "instructor_id",
                                    "is_online",
                                    "schedule_id"
                                ],
                                "properties": {
                                    "course_name": {
                                        "description": "Name of the course",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Mathematics"
                                    },
                                    "instructor_id": {
                                        "description": "ID of the instructor teaching the course",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "is_online": {
                                        "description": "Whether the course is online or in-person",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "schedule_id": {
                                        "description": "ID of the schedule associated with the course",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Course created successfully"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/course/{id}": {
            "put": {
                "tags": [
                    "Course Management"
                ],
                "summary": "Update a Course",
                "description": "Update details of an existing course.",
                "operationId": "b318336f8992b3e357d97a2ba400f4d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_name",
                                    "instructor_id",
                                    "is_online",
                                    "schedule_id"
                                ],
                                "properties": {
                                    "course_name": {
                                        "description": "Name of the course",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Advanced Mathematics"
                                    },
                                    "instructor_id": {
                                        "description": "ID of the instructor teaching the course",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "is_online": {
                                        "description": "Whether the course is online or in-person",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "schedule_id": {
                                        "description": "ID of the schedule associated with the course",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Course updated successfully"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Course Management"
                ],
                "summary": "Delete a Course",
                "description": "Delete an existing course.",
                "operationId": "53993bf6c65af86b51cc51ab254a0d7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Course deleted successfully"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}